[Chore #157694748] Fix pushover integration tests
